+++ app/src/main/java/com/example/umc_hackathon/FormListActivity.kt
@@ -1,0 +1,36 @@
+package com.example.umc_hackathon
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.umc_hackathon.databinding.ActivityFormListBinding
+
+class FormListActivity : AppCompatActivity() {
+
+    val TAG: String = "<FormListActivity>"
+    var modelList = ArrayList<MySurvey>()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val binding = ActivityFormListBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        // 리스트 생성
+        for (i in 1..10){
+            val mySurvey = MySurvey(title = "사회현상에 대한 소비자 인식 $i")
+            this.modelList.add(mySurvey)
+        }
+
+        binding.rvListItem.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
+        binding.rvListItem.setHasFixedSize(true)
+        binding.rvListItem.adapter = MyRecyclerAdapter(modelList)
+
+        // 이벤트 리스너
+        binding.mainButton.setOnClickListener {
+            val intent = Intent(this, FormCreateActivity::class.java)
+            startActivity(intent)
+        }
+    }
+
+}
