+++ app/src/main/res/layout/activity_form_detail.xml
@@ -34,7 +34,7 @@
             android:textSize="20sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
-            
+
         <TextView
             android:id="@+id/survey_day_text"
             android:layout_width="wrap_content"
